% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acyclic_graph.R
\name{acyclic_graph_from_df}
\alias{acyclic_graph_from_df}
\title{From data.frame}
\usage{
acyclic_graph_from_df(df)
}
\arguments{
\item{df}{The data.frame to create the graph from. It must have two columns,
\code{parent} and \code{child}.}
}
\value{
A new graph.
}
\description{
Creates a graph from a data.frame.

The data.frame must have two columns, one for the parent and one for the
child.
}
