% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{AcyclicGraph}
\alias{AcyclicGraph}
\alias{$.AcyclicGraph}
\title{Structure that holds a graph with no cycles. You can create
nodes and query the graph for parents, children, leaves and
least common parents.}
\format{
An object of class \code{environment} of length 10.
}
\usage{
AcyclicGraph
}
\description{
Structure that holds a graph with no cycles. You can create
nodes and query the graph for parents, children, leaves and
least common parents.
}
\keyword{datasets}
