% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_child}
\alias{add_child}
\title{Add Child}
\usage{
add_child(graph, parent, child)
}
\arguments{
\item{graph}{The graph to add the child to.}

\item{parent}{The ID of the parent node.}

\item{child}{The ID of the child node.}
}
\value{
A reference to the graph passed in.
}
\description{
Adds a child to a node. If the child already exists, nothing happens.

If the parent or the child do not exist, they are created.

\strong{Note}: This function modifies the graph in place. It does not return a
new graph. This is done for performance and memory reasons. It works in a
similar way to the \code{data.table} package.
}
