% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edge.R
\name{add_edge}
\alias{add_edge}
\title{Add an edge to a graph}
\usage{
add_edge(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{The id of the \code{from} node.}

\item{to}{The id of the \code{to} node.}
}
\value{
The updated graph object
}
\description{
Adds an edge from one node to another in a
a directed graph.

NOTE: Not all graphs are able to be modified.
}
