% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node.R
\name{add_node}
\alias{add_node}
\title{Add a new node to a graph}
\usage{
add_node(graph, node_id, data = NULL)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The unique id of the node to be added}

\item{data}{R object to be used as the data for the node}
}
\value{
The updated graph object
}
\description{
Appends a new node to the graph in-place with
some arbitrary data.

NOTE: Not all graphs are able to be modified.
}
