% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_path}
\alias{add_path}
\title{Add a path to a graph}
\usage{
add_path(graph, path)
}
\arguments{
\item{graph}{A graph object}

\item{path}{A character vector that describes the path}
}
\value{
The updated graph object
}
\description{
Adds all of the edges that make up the given
path to the graph.

NOTE: Not all graphs are able to be modified.
}
