% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_graph.data.frame}
\alias{as_graph.data.frame}
\title{Data.frame as Graph}
\usage{
\method{as_graph}{data.frame}(x, type, ...)
}
\arguments{
\item{x}{The data.frame to convert to a graph.}

\item{type}{The type of graph to convert to. Currently only \code{acyclic} is
supported.}

\item{...}{Ignored.}
}
\value{
A graph of the given type.
}
\description{
Converts a data.frame to a graph.
}
