% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_graph}
\alias{as_graph}
\title{As Graph}
\usage{
as_graph(x, type, ...)
}
\arguments{
\item{x}{The object to convert to a graph.}

\item{type}{The type of graph to convert to. Currently only \code{acyclic} is
supported.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A graph of the given type.
}
\description{
Attempts to convert the object to a graph.
}
