% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{build_directed}
\alias{build_directed}
\title{Builder into a Directed Graph}
\usage{
build_directed(graph_builder)
}
\arguments{
\item{graph_builder}{A graph builder object}
}
\value{
A DirectedGraph Object
}
\description{
Builds a graph builder into a new DirectedGraph
object.

NOTE: This will consume the builder. It will leave an empty
builder in its place.
}
