% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\title{Get the children on a node}
\usage{
children(graph, nodes)
}
\arguments{
\item{graph}{A graph object}

\item{nodes}{A character vector of nodes to find children for}
}
\value{
A character vector
}
\description{
Get a list of the node ids of the children of
the provided node.
}
