% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_edges.R
\name{clear_edges}
\alias{clear_edges}
\title{Remove all edges from a graph}
\usage{
clear_edges(graph)
}
\arguments{
\item{graph}{A graph object}
}
\value{
The updated graph object
}
\description{
Removes all edges from a graph in-place.

NOTE: Not all graphs are able to be modified.
}
