% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{clone_graph}
\alias{clone_graph}
\title{Clone Graph}
\usage{
clone_graph(graph)
}
\arguments{
\item{graph}{The graph to clone.}
}
\value{
A new graph that is a copy of the original.
}
\description{
Creates a copy of the graph.
}
