% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_clone.R
\name{deep_clone}
\alias{deep_clone}
\title{Get a deep clone of a graph}
\usage{
deep_clone(graph)
}
\arguments{
\item{graph}{A graph object}
}
\value{
A new deep copy of the graph
}
\description{
Creates a new copy of a graph.
}
