% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_exists.R
\name{edge_exists}
\alias{edge_exists}
\title{Validate if an edge exists}
\usage{
edge_exists(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{ID of the \code{from} node}

\item{to}{ID of the \code{to} node}
}
\value{
A logical describing if the edge exists
}
\description{
Returns \code{TRUE} if the edge exists.
}
