% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{filter_tree}
\alias{filter_tree}
\title{Filters graph with selected children}
\usage{
filter_tree(x, selected)
}
\arguments{
\item{x}{The graph to search in.}

\item{selected}{Children selected for filter.}
}
\value{
A graph with the selected children.
}
\description{
Filters graph with selected children
}
