% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_path.R
\name{find_all_paths}
\alias{find_all_paths}
\title{Find all paths between two nodes}
\usage{
find_all_paths(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{The starting node of the path}

\item{to}{The ending node of the path}
}
\value{
A list of character vectors
}
\description{
Find all the paths between two nodes in a graph.

Not all graphs support this function. Currently only
\code{DirectedAcyclicGraph} supports this.
}
