% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_least_common_parents}
\alias{find_least_common_parents}
\title{Get Least Common Parents from an Acyclic Graph}
\usage{
find_least_common_parents(graph, nodes)
}
\arguments{
\item{graph}{The graph to get the least common parents from.}

\item{nodes}{The nodes to get the least common parents of.}
}
\value{
A character vector of the least common parents of the nodes.
}
\description{
Gets the least common parents of a set of nodes.
This is the set of parents that are parents of all the nodes and that have
been selected.

This is useful for example if you want to group by the set of parents of a
set of nodes.
}
