% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_leaves}
\alias{find_leaves}
\title{Get Leaves / Maximum Depth}
\usage{
find_leaves(graph, node)
}
\arguments{
\item{graph}{The graph to get the leaves from.}

\item{node}{The ID of the node to get the leaves of.}
}
\value{
A character vector of the leaves of the node.
}
\description{
Gets the leaves of the graph that descend from a node.
}
