% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_path.R
\name{find_path_one_to_many}
\alias{find_path_one_to_many}
\title{Find the shortest path from one node to many}
\usage{
find_path_one_to_many(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{The starting node of the path}

\item{to}{A character vector of nodes}
}
\value{
A list of paths
}
\description{
Find the shortest path from one node to many

Not all graphs support this function. Currently only
\code{DirectedAcyclicGraph} supports this.
}
\examples{
edges <- data.frame(
  parent = c("A", "A", "B", "Z"),
  child =  c("B", "Z", "Z", "F")
)

graph <- graph_builder() |>
  populate_edges(edges, parent, child) |>
  build_acyclic()

find_path_one_to_many(graph, "A", edges$child)
}
