% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_roots}
\alias{find_roots}
\title{Find Roots}
\usage{
find_roots(graph)
}
\arguments{
\item{graph}{The graph to get the roots from.}
}
\value{
A character vector of the roots of the graph.
}
\description{
Gets the roots of the graph.
}
