% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roots.R
\name{get_all_roots}
\alias{get_all_roots}
\title{Get the all the root nodes of a graph}
\usage{
get_all_roots(graph, ...)
}
\arguments{
\item{graph}{A graph object}

\item{...}{Unused}
}
\value{
A character vector of nodes
}
\description{
Retrieves the nodes in a graph that have no parents
}
