% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{get_children}
\alias{get_children}
\title{Get the children on a node}
\usage{
get_children(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The id of the node}
}
\value{
A character vector
}
\description{
Get a list of the node ids of the children of
the provided node.
}
