% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaves.R
\name{get_leaves}
\alias{get_leaves}
\title{Get the leave nodes of a graph}
\usage{
get_leaves(graph, node_ids)
}
\arguments{
\item{graph}{A graph object}

\item{node_ids}{The node ids of which the leaves must be descendants of}
}
\value{
A character vector of node ids
}
\description{
Retrieves the nodes in a graph that have no children.

If no \code{node_ids} argument is supplied then all of
the leaves in the graph are returned.
}
