% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaves_as_df.R
\name{get_leaves_as_df}
\alias{get_leaves_as_df}
\title{Get leaves as a data frame}
\usage{
get_leaves_as_df(graph, nodes)
}
\arguments{
\item{graph}{A graph object}

\item{nodes}{A character vector of node IDs}
}
\value{
A data frame of leaves
}
\description{
Get leaves of a set of nodes in a data frame format.
}
