% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{get_node}
\alias{get_node}
\title{Get the node from a graph object}
\usage{
get_node(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The unique id of the node to be retrieved}
}
\value{
A node object
}
\description{
Retrives the a specific node from a graph with
its data.
}
