% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{get_node_data}
\alias{get_node_data}
\title{Retrieve the data out of a node}
\usage{
get_node_data(node)
}
\arguments{
\item{node}{The node to retrieve the data from}
}
\value{
R object in the node
}
\description{
Gets the data from a node object
}
