% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{get_node_ids}
\alias{get_node_ids}
\title{Get the ids of a group of nodes from a graph object}
\usage{
get_node_ids(graph)
}
\arguments{
\item{graph}{A graph object}
}
\value{
A character vector containing the node ids
}
\description{
Retrieves a character vector with the ids of nodes
in a graph
}
