% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{get_nodes}
\alias{get_nodes}
\title{Get a group of nodes from a graph object}
\usage{
get_nodes(graph, node_ids)
}
\arguments{
\item{graph}{A graph object}

\item{node_ids}{The unique ids of the nodes to be retrieved}
}
\value{
A list of node objects
}
\description{
Retrieves a list of nodes with their data from
a graph object.

If no \code{node_ids} argument is supplied then all of
the nodes in the graph a returned in a list.
}
