% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{get_parents}
\alias{get_parents}
\title{Get the parents on a node}
\usage{
get_parents(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The id of the node}
}
\value{
A character vector
}
\description{
Get a list of the node ids of the parents of
the provided node.
}
