% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roots.R
\name{get_roots}
\alias{get_roots}
\title{Get the root nodes of a graph}
\usage{
get_roots(graph, node_ids)
}
\arguments{
\item{graph}{A graph object}

\item{node_ids}{The node ids of which the roots must be descendants of}
}
\value{
A character vector of node ids
}
\description{
Retrieves the nodes in a graph that have no parents

If no \code{node_ids} argument is supplied then all of
the roots in the graph are returned.
}
