% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roots.R
\name{get_roots_over}
\alias{get_roots_over}
\title{Get the root nodes of a graph over some nodes}
\usage{
get_roots_over(graph, nodes)
}
\arguments{
\item{graph}{A graph object}

\item{nodes}{A character vector of nodes to find roots for}
}
\value{
A character vector of nodes
}
\description{
Retrieves the nodes in a graph that have no parents
over a certain node or group of nodes
}
