% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{graph_builder}
\alias{graph_builder}
\title{A new builder for a graph based on the type}
\usage{
graph_builder(type = "directed")
}
\arguments{
\item{type}{The type of graph}
}
\description{
Object used to build graphs
}
