% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{graph_from_json}
\alias{graph_from_json}
\title{Get a graph from a JSON string}
\usage{
graph_from_json(text, type = c("directed", "dag"))
}
\arguments{
\item{text}{The JSON string to be deserialized}

\item{type}{The type of graph the JSON represents}
}
\value{
A graph object
}
\description{
Get a graph from a JSON string
}
