% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{graph_to_bin}
\alias{graph_to_bin}
\title{Save the graph into a binary blob}
\usage{
graph_to_bin(graph, path)
}
\arguments{
\item{graph}{A graph object}

\item{path}{Path to a file to save the graph into}
}
\value{
Run for its side-effects
}
\description{
Save the graph into a binary blob
}
\examples{
graph <- graph_builder() |>
  add_edge("A", "B") |>
  build_directed()

graph_to_bin(graph)
}
\seealso{
Other graphs i/o: 
\code{\link{graph_from_bin}()}
}
\concept{graphs i/o}
