% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{graph_to_json}
\alias{graph_to_json}
\title{Serialize the graph into a JSON string}
\usage{
graph_to_json(graph, pretty = FALSE)
}
\arguments{
\item{graph}{A graph object}

\item{pretty}{If the JSON string should be prettified}
}
\value{
A JSON string
}
\description{
Serialize the graph into a JSON string
}
