% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_children.R
\name{has_children}
\alias{has_children}
\title{Checks if a node in a graph has children}
\usage{
has_children(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The id of the node to check}
}
\value{
A logical describing if the node has children
}
\description{
This function validates if the node has an edge pointing
to any other node.
}
