% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_parents.R
\name{has_parents}
\alias{has_parents}
\title{Checks if a node in a graph has parents}
\usage{
has_parents(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The id of the node to check}
}
\value{
A logical describing if the node has parents
}
\description{
This function validates if any edge points to the
given node.
}
