% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{into_dag}
\alias{into_dag}
\title{Convert a Directed Graph into a Directed Acyclic Graph}
\usage{
into_dag(directed_graph)
}
\arguments{
\item{directed_graph}{A Directed Graph object}
}
\value{
A new DirectedAcyclicGraph object
}
\description{
Validates and converts a Directed Graph into an Acyclic Directed Graph
applying a topological sort.
}
