% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{into_directed}
\alias{into_directed}
\title{Convert an Directed Acyclic Graph into a Directed Graph}
\usage{
into_directed(directed_acyclic_graph)
}
\arguments{
\item{directed_acyclic_graph}{A Directed Acyclic Graph object}
}
\value{
A new DirectedGraph object
}
\description{
Transforms a DirectedAcyclicGraph object into a DirectedGraph object.

This is useful because the structure of a DirectedAcyclicGraph is
immutable. In order to modify the nodes or edges of a
DirectedAcyclicGraph requires converting it to a DirectedGraph first.
}
