% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{new_directed_graph}
\alias{new_directed_graph}
\title{Initialize a New Directed Graph}
\usage{
new_directed_graph(type)
}
\value{
A new Directed Graph object.
}
\description{
Initializes a new and empty Directed Graph.
}
