% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{new_graph}
\alias{new_graph}
\title{Initialize a New Graph}
\usage{
new_graph(type)
}
\arguments{
\item{type}{The type of graph to create. Currently only \code{acyclic} is
supported.}
}
\value{
A new graph of the given type.
}
\description{
Initializes a new graph with the given type.
}
