% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{orbweaver-package}
\alias{orbweaver}
\alias{orbweaver-package}
\title{orbweaver: Fast and Efficient Graph Data Structures}
\description{
Empower your data analysis with 'orbweaver', an R package designed for effortless construction and analysis of graph data structures. With 'orbweaver', you can seamlessly build and manipulate graph structures, leveraging its high-performance methods for filtering, joining, and mutating data within the R environment. Drawing inspiration from the efficiency of the 'data.table' package, 'orbweaver' ensures that mutations and changes to the graph are performed in place, streamlining your workflow for optimal productivity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ixpantia/orbweaver-r}
  \item Report bugs at \url{https://github.com/ixpantia/orbweaver-r/issues}
}

}
\author{
\strong{Maintainer}: Andres Quintero \email{andres@ixpantia.com}

Other contributors:
\itemize{
  \item ixpantia, SRL \email{hola@ixpantia.com} [copyright holder]
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
