% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building.R
\name{populate_nodes}
\alias{populate_nodes}
\title{Populates the nodes of a graph from a \code{tibble}}
\usage{
populate_nodes(graph, nodes_df, node_id_col, data_col = NULL)
}
\arguments{
\item{graph}{A graph object}

\item{nodes_df}{A \code{tibble} with a node id and data column}

\item{node_id_col}{The name of the column containing the node ids}

\item{data_col}{The name of the column containing the data of the nodes}
}
\value{
The updated graph object
}
\description{
Adds a set of nodes from a \code{tibble} to a graph
}
