% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_edge.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove an edge from a graph}
\usage{
remove_edge(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{The \code{from} node id of the edge to be removed}

\item{to}{The \code{to} node id of the edge to be removed}
}
\value{
The updated graph object
}
\description{
Removes an edge from the graph in-place.

NOTE: Not all graphs are able to be modified.
}
