% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_node.R
\name{remove_node}
\alias{remove_node}
\title{Remove a node from a graph}
\usage{
remove_node(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The id of the node to be removed}
}
\value{
The updated graph object
}
\description{
Removes a node from the graph in-place.

NOTE: Not all graphs are able to be modified.
}
