% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{search_for_node}
\alias{search_for_node}
\title{Search for a node inside a graph by name}
\usage{
search_for_node(graph, node_id, case_sensitive = TRUE)
}
\arguments{
\item{graph}{The graph to search in.}

\item{node_id}{The string to search for.}

\item{case_sensitive}{Whether the search should be case sensitive.}
}
\value{
A character vector of the nodes that match the search.
}
\description{
Searches for nodes with a name that matches the given string.
}
