% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{set_node_data}
\alias{set_node_data}
\title{Update the data of a node}
\usage{
set_node_data(node, data)
}
\arguments{
\item{node}{The node to set the data into}

\item{data}{An arbitrary R object}
}
\value{
The updated node object
}
\description{
Updates the data of a node in-place
}
