% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_graph}
\alias{subset_graph}
\title{Get a subset or a graph}
\usage{
subset_graph(graph, node_id)
}
\arguments{
\item{graph}{A graph object}

\item{node_id}{The node id from which to split the tree}
}
\value{
A new graph object
}
\description{
todo!
}
