% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node.R
\name{update_node_data}
\alias{update_node_data}
\title{Update the data of a node}
\usage{
update_node_data(graph, node_id, data)
}
\arguments{
\item{data}{An arbitrary R object}

\item{node}{The node to set the data into}
}
\value{
The updated node object
}
\description{
Updates the data of a node in-place
}
