% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_edge}
\alias{add_edge}
\title{Add an edge to a graph builder}
\usage{
add_edge(graph_builder, from, to)
}
\arguments{
\item{graph_builder}{A graph builder_object}

\item{from}{The \code{from} node.}

\item{to}{The \code{to} node.}
}
\value{
The updated graph builder object
}
\description{
Adds an edge from one node to another in a
a directed graph builder.
}
