% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_node}
\alias{add_node}
\title{Add Node}
\usage{
add_node(graph, node)
}
\arguments{
\item{graph}{The graph to add the node to.}

\item{node}{The ID of the node to add.}
}
\value{
A reference to the graph passed in.
}
\description{
Adds a node to the graph. If the node already exists, nothing happens.

\strong{Note}: This function modifies the graph in place. It does not return a
new graph. This is done for performance and memory reasons. It works in a
similar way to the \code{data.table} package.
}
