% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{add_path}
\alias{add_path}
\title{Add a path to a graph}
\usage{
add_path(graph_builder, path)
}
\arguments{
\item{graph_builder}{A graph builder_object}

\item{path}{A character vector that describes the path}
}
\value{
The updated graph builder object
}
\description{
Adds all of the edges that make up the given
path to the graph.
}
\examples{
graph_builder() |>
  add_path(c("A", "B", "C"))
}
\seealso{
Other build graphs: 
\code{\link{add_edge}()},
\code{\link{build_acyclic}()},
\code{\link{build_directed}()},
\code{\link{graph_builder}()},
\code{\link{populate_edges}()}
}
\concept{build graphs}
