% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as.list.AcyclicGraph}
\alias{as.list.AcyclicGraph}
\title{Convert to List}
\usage{
\method{as.list}{AcyclicGraph}(x, ...)
}
\arguments{
\item{x}{The graph to convert to a list.}

\item{...}{Ignored.}
}
\value{
A list representation of the graph.
}
\description{
Converts the graph to a list.
}
