% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_all_paths}
\alias{find_all_paths}
\title{Find all paths between two nodes}
\usage{
find_all_paths(graph, from, to)
}
\arguments{
\item{graph}{The graph to search in.}

\item{from}{The ID of the node to start the search from.}

\item{to}{The ID of the node to end the search at.}
}
\value{
A list of character vectors of the paths between the two nodes.
}
\description{
Finds all paths between two nodes.
}
