% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_children}
\alias{get_children}
\title{Get Children}
\usage{
get_children(graph, node)
}
\arguments{
\item{graph}{The graph to get the children from.}

\item{node}{The ID of the node to get the children of.}
}
\value{
A character vector of the children of the node.
}
\description{
Gets the children of a node.
}
