% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaves.R
\name{get_leaves_under}
\alias{get_leaves_under}
\title{Get the leaf nodes of a graph under some nodes}
\usage{
get_leaves_under(graph, nodes)
}
\arguments{
\item{graph}{A graph object}

\item{nodes}{A character vector of nodes to find leaves for}
}
\value{
A character vector of nodes
}
\description{
Retrieves the nodes in a graph that have no children
under a certain node or group of nodes
}
