% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_parents}
\alias{get_parents}
\title{Get Parents}
\usage{
get_parents(graph, node)
}
\arguments{
\item{graph}{The graph to get the parents from.}

\item{node}{The ID of the node to get the parents of.}
}
\value{
A character vector of the parents of the node.
}
\description{
Gets the parents of a node.
}
