% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_parents.R
\name{has_parents}
\alias{has_parents}
\title{Checks if a node in a graph has parents}
\usage{
has_parents(graph, nodes)
}
\arguments{
\item{graph}{A graph object}

\item{nodes}{A character vector of nodes to determine}
}
\value{
A logical vector with the same length as \code{nodes}
}
\description{
This function validates if any edge points to the
given node.
}
\examples{
graph <- graph_builder() |>
  add_edge(from = "A", to = "B") |>
  build_directed()
graph

graph |> has_parents(nodes = "A")
graph |> has_parents(nodes = "B")
}
