% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{populate_edges}
\alias{populate_edges}
\title{Populates the edges of a graph from a \code{tibble}}
\usage{
populate_edges(graph_builder, edges_df, parent_col, child_col)
}
\arguments{
\item{graph_builder}{A graph builder object}

\item{edges_df}{A \code{tibble} with a parent and child variable}

\item{parent_col}{The name of the column containing the parents}

\item{child_col}{The name of the column containing the children}
}
\value{
The updated graph builder object
}
\description{
Adds a set of edges from a \code{tibble} to a graph
}
